requirements.txt
astroid==2.2.5
blessings==1.7
bpython==0.17.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
curtsies==0.3.0
Flask==1.1.1
fpdf==1.7.2
greenlet==0.4.15
idna==2.8
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.3.1
lxml==4.3.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-lang==0.5.0
numpy==1.16.4
pep8==1.7.1
Pillow==5.4.1
py==1.7.0
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
PySocks==1.6.8
pyTelegramBotAPI==3.6.6
python-docx==0.8.10
reportlab==3.5.13
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
typing==3.6.6
urllib3==1.24.1
virtualenv==16.7.8
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.1
